* fixing package paths after package renames * fixed typo
