+++ src/main/scala/za/co/absa/fadb/naming_conventions/SnakeCaseNaming.scala
@@ -13,10 +13,10 @@
  * limitations under the License.
  */
 
-package za.co.absa.faDB.namingConventions
+package za.co.absa.fadb.naming_conventions
 
-import za.co.absa.faDB.namingConventions.lettersCase.LettersCase
-import za.co.absa.faDB.namingConventions.lettersCase.LettersCase.LowerCase
+import za.co.absa.fadb.naming_conventions.lettersCase.LettersCase
+import za.co.absa.fadb.naming_conventions.lettersCase.LettersCase.LowerCase
 
 class SnakeCaseNaming(lettersCase: LettersCase) extends NamingConvention {
   private def camelCaseToSnakeCase(s: String): String = {
