+++ src/test/scala/za/co/absa/fadb/naming_conventions/ExplicitNamingRequiredSuite.scala
@@ -1,8 +1,8 @@
-package za.co.absa.faDB.namingConventions
+package za.co.absa.fadb.naming_conventions
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import za.co.absa.faDB.exceptions.NamingException
+import za.co.absa.fadb.exceptions.NamingException
 
 class ExplicitNamingRequiredSuite extends AnyWordSpec with Matchers {
   private val explicitNamingRequired = new ExplicitNamingRequired()
