+++ src/test/scala/za/co/absa/fadb/naming_conventions/SnakeCaseNamingSuite.scala
@@ -1,8 +1,8 @@
-package za.co.absa.faDB.namingConventions
+package za.co.absa.fadb.naming_conventions
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import za.co.absa.faDB.namingConventions.lettersCase.LettersCase._
+import za.co.absa.fadb.naming_conventions.lettersCase.LettersCase._
 
 class SnakeCaseNamingSuite extends AnyWordSpec with Matchers {
   private class ThisIsATestClass
